corrected typos
