+++ routes/api/user-routes.js
@@ -1,5 +1,5 @@
-const router = require("express").Router();
-const { User, Test, Comment } = require("../../models");
+const router = require('express').Router();
+const { User, Test, Comment } = require('../../models');
 const withAuth = require('../../utils/auth');
 
 // GET /api/users
@@ -76,31 +76,33 @@ router.post("/", (req, res) => {
 
 router.post('/login', (req, res) => {
     User.findOne({
-        where: {
-            username: req.body.username
-        }
+      where: {
+        username: req.body.username
+      }
     }).then(dbUserData => {
-        if (!dbUserData) {
-            res.status(400).json({ message: 'No user with that username!' });
-            return;
-        }
-
-        const validPassword = dbUserData.checkPassword(req.body.password);
-
-        if (!validPassword) {
-            res.status(400).json({ message: 'Incorrect password!' });
-            return;
-        }
-
-        req.session.save(() => {
-            req.session.user_id = dbUserData;
-            req.session.username = dbUserData.username;
-            req.session.loggedIn = true;
-
-            res.json({ user: dbUserData, message: 'You are now logged in!' });
-        })
+      if (!dbUserData) {
+        res.status(400).json({ message: 'No user with that username!' });
+        return;
+      }
+  
+      const validPassword = dbUserData.checkPassword(req.body.password);
+  
+      if (!validPassword) {
+        res.status(400).json({ message: 'Incorrect password!' });
+        return;
+      }
+  
+      req.session.save(() => {
+        // declare session variables
+        req.session.user_id = dbUserData.id;
+        req.session.username = dbUserData.username;
+        
+        req.session.loggedIn = true;
+  
+        res.json({ user: dbUserData, message: 'You are now logged in!' });
+      });
     });
-});
+  });
 
 router.post('/logout', (req, res) => {
     if (req.session.loggedIn) {
