+++ views/login.handlebars
@@ -1,5 +1,5 @@
 <div class="column">
-<h2 class="title pb-4">Login:</h2><hr>
+<h2 class="title pb-4">login:</h2><hr>
 <form class="login-form">
   <div>
     <label for="username-login">username:</label>
@@ -10,8 +10,8 @@
     <input type="password" id="password-login" />
   </div>
   <div>
-    <button type="submit" class="button is-primary">Login</button>
-    <a href="/signup"><button type="button" class="button is-danger">Signup</button></a>
+    <button type="submit" class="button is-primary">login</button>
+    <a href="/signup"><button type="button" class="button is-danger">signup</button></a>
   </div>
 </form>
 </div>
