+++ views/signup.handlebars
@@ -1,5 +1,5 @@
 <div class="column">
-<h2 class="title pb-4">Create new account:</h2><hr>
+<h2 class="title pb-4">create new account:</h2><hr>
 <form class="signup-form">
   <div>
     <label for="username-signup">username:</label>
@@ -10,7 +10,7 @@
     <input type="password" id="password-signup" />
   </div>
   <div>
-    <button type="submit" class="button is-primary">Signup</button>
+    <button type="submit" class="button is-primary">signup</button>
   </div>
 </form>
 </div>
